requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
arize-phoenix==7.9.2
arize-phoenix-evals==0.19.0
arize-phoenix-otel==0.7.1
async-timeout==4.0.3
attrs==25.1.0
Authlib==1.4.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
fastapi==0.115.7
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
googleapis-common-protos==1.66.0
graphql-core==3.2.6
greenlet==3.1.1
groq==0.15.0
grpc-interceptor==0.15.4
grpcio==1.70.0
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-groq==0.2.3
langchain-text-splitters==0.3.5
langsmith==0.3.2
llvmlite==0.43.0
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.26.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
openinference-instrumentation==0.1.21
openinference-semantic-conventions==0.1.12
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==5.24.1
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlean.py==3.47.0
starlette==0.45.3
strawberry-graphql==0.253.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
